fix(review): address memory handoff review feedback
